towards-targets: Move equation module inside finite_differences
