+++ devito/dse/rewriters.py
@@ -4,13 +4,13 @@ from time import time
 
 from sympy import cos, sin
 
-from devito.equation import Eq
 from devito.ir import (DataSpace, IterationSpace, Interval, IntervalGroup, Cluster,
                        detect_accesses, build_intervals)
 from devito.dse.aliases import collect
 from devito.dse.manipulation import (collect_nested, common_subexprs_elimination,
                                      make_is_time_invariant)
 from devito.exceptions import DSEException
+from devito.finite_differences import Eq
 from devito.logger import dse_warning as warning
 from devito.symbolics import (bhaskara_cos, bhaskara_sin, estimate_cost, freeze,
                               pow_to_mul, q_leaf, q_sum_of_product, q_terminalop,
