+++ devito/finite_differences/__init__.py
@@ -1,6 +1,7 @@
-from devito.finite_differences.differentiable import *  # noqa
-from devito.finite_differences.finite_difference import *  # noqa
-from devito.finite_differences.derivative import *  # noqa
-from devito.finite_differences.tools import generate_fd_shortcuts  # noqa
-from devito.finite_differences.coefficients import *  # noqa
-from devito.finite_differences.operators import *  # noqa
+from .differentiable import *  # noqa
+from .finite_difference import *  # noqa
+from .derivative import *  # noqa
+from .tools import generate_fd_shortcuts  # noqa
+from .coefficients import *  # noqa
+from .operators import *  # noqa
+from .equation import *  # noqa
