+++ devito/types/sparse.py
@@ -5,8 +5,7 @@ import sympy
 import numpy as np
 from cached_property import cached_property
 
-from devito.equation import Eq, Inc
-from devito.finite_differences import Differentiable, generate_fd_shortcuts
+from devito.finite_differences import Differentiable, Eq, Inc, generate_fd_shortcuts
 from devito.logger import warning
 from devito.mpi import MPI, SparseDistributor
 from devito.symbolics import INT, cast_mapper, indexify, retrieve_function_carriers
